refactor(voice): split the lobby card out of voice_view.dart
This PR's pre-join card expansion pushed voice_view.dart to 1099 lines,
crossing the 1000-line file-size guideline. Extract VoiceLobbyBody (plus
its private _PreJoinChip/_LobbyAvatar helpers) into voice_lobby.dart — a
self-contained "who's here + join" unit — bringing voice_view.dart back
to 771 lines. No behavior change.
